django/requirements.txt
Django
djangorestframework
django-cors-headers
psycopg2
djangorestframework-jwt
cryptography
python-jose